moving modulemanager to managers folder added configmanager, still has to be functional
