+++ src/main/kotlin/org/xodium/vanillaplus/managers/ConfigManager.kt
@@ -1,0 +1,7 @@
+package org.xodium.vanillaplus.managers
+
+object ConfigManager {
+    init {
+        // TODO: implement.
+    }
+}
